Core/PacketIO: Disabled SMSG_DB_REPLY

+++ src/server/game/Server/Packets/QueryPackets.cpp
@@ -258,9 +258,9 @@ void WorldPackets::Query::DBQueryBulk::Read()
 
 WorldPacket const* WorldPackets::Query::DBReply::Write()
 {
-    _worldPacket << TableHash;
-    _worldPacket << RecordID;
-    _worldPacket << Timestamp;
+    _worldPacket << uint32(TableHash);
+    _worldPacket << uint32(RecordID);
+    _worldPacket << uint32(Timestamp);
     _worldPacket.WriteBit(Allow);
     _worldPacket << uint32(Data.size());
     _worldPacket.append(Data);
